fixed a warning
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/perso/riviere/qroll/trunk@59773 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/libroll/gui/scoreBox.cc
@@ -229,7 +229,7 @@ void QRScoreBox::changeScore()
 
   for( ib=pl.bombs.begin(); ib!=eb; ++ib )
   {
-    if( ib->second != _bombs[ ib->first ]->intValue() )
+    if( (int) ib->second != _bombs[ ib->first ]->intValue() )
       setBombs( ib->first, ib->second );
   }
   for( i=0; i<8; ++i )
